BancodeDados/Exercicios/ServidorEx/requirements.txt
Flask>=2.2
psycopg2-binary>=2.9
sqlalchemy>=1.4
flask-restful>=0.3.9
python-dotenv>=1.0
pandas>=2.0
nbformat>=5.0